fix fibo 4

+++ functions_nested_loops/104-fibonacci.c
@@ -8,16 +8,16 @@
 
 int main(void)
 {
-	double a = 1;
-	double b = 2;
-	double c;
+	float a = 1;
+	float b = 2;
+	float c;
 	int i;
 
 	printf("1, 2, ");
 	for (i = 0; i < 96; i++)
 	{
 		c = a + b;
-		printf("%.0lf", c);
+		printf("%.0f", c);
 		if (i < 95)
 			printf(", ");
 		a = b;
